move ThreadSafeDefaultDict
